타입 일치화

+++ src/main/java/step1/Customer.java
@@ -51,7 +51,7 @@ class Customer {
 	private double amountFor(Rental each) {
 		// determine amounts for each line
 
-		int thisAmount = 0;
+		double thisAmount = 0;
 		switch (each.getMovie().getPriceCode()) {
         case Movie.REGULAR:
             thisAmount += 2;
